feat: copy directly to clipboard, add copy content option
Moves the options in the "Chat 2" submenu when right clicking a message
to the parent menu if there are no specific context menu items for the
right clicked payload. This moves the "Copy" button to the main context
menu in the majority of cases and reduces the amount of clicks required
to copy.

Changes the "Copy" button to copy directly to the system clipboard and
adds a notification.

Adds a "Copy content" button beneath the "Copy" button which only copies
the message content (not the sender). If the message doesn't have a
sender, this option does not appear as they function identically in such
cases.